Refactor components outside
