+++ src/firstYearGuide/components/assets/css/components/road.css
@@ -1,0 +1,16 @@
+.road-container {
+    position: absolute;
+    bottom: 0;
+    left: 0;
+
+    width: 100%;
+    height: 100%;
+
+    background-image: url("../../images/components/mobile-road.png");
+    background-repeat: repeat-x;
+    background-size: auto 100%;
+
+    overflow: hidden;
+
+    z-index: 2;
+}
