+++ src/firstYearGuide/components/assets/css/components/stars.css
