+++ src/firstYearGuide/components/assets/images/components/mobile-stars.png
