+++ src/firstYearGuide/components/assets/images/components/mobile-stars@2x.png
