+++ src/firstYearGuide/components/assets/images/components/mobile-tree-night@2x.png
