+++ src/firstYearGuide/components/road/road.jsx
@@ -1,0 +1,7 @@
+import '../assets/css/components/road.css';
+
+const Road = () => (
+  <div className='road-container'/>
+);
+
+export default Road;
