+++ src/firstYearGuide/components/scrollButton/scrollButton.jsx
@@ -1,4 +1,4 @@
-import "../../assets/css/components/scrollButton.css";
+import "../assets/css/components/scrollButton.css";
 
 const ScrollButton = (props) => {
     let floatDirection = "left";
