+++ src/firstYearGuide/desktop/assets/css/desktopView.css
@@ -1,0 +1,15 @@
+.desktopView{
+    width: 100vw;
+    height: 100svh;
+    background-color: black;
+}
+
+.desktopView__buildingWrapper{
+    width: 100%;
+    height: 100%;
+
+    display: flex;
+    align-items: flex-end;
+    padding-bottom: 10vh;
+    justify-content: center;
+}
