+++ src/firstYearGuide/desktop/desktopView.jsx
@@ -1,18 +1,21 @@
-import desktopImg from "./assets/images/FYG_Desktop.png";
+import buildingCluster from "./assets/images/building_Cluster.png";
+import "./assets/css/desktopView.css";
 
 const DesktopView = (props) => {
     return(
-        <div style={{
-            width: "100vw",
-            height: "100svh",
-        }}>
-            <img 
-                src={desktopImg}
-                style={{
-                    width: "100vw",
-                    height: "100svh"
-                }}
-            />
+        <div className="desktopView">
+            <div className="desktopView__buildingWrapper">
+                <img 
+                    src={buildingCluster}
+                    style={{
+                        width: "100vw",
+                        maxWidth: '900px',
+                    }}
+                />
+            </div>
+            
+
+
         </div>
     )
 }
