--- src/firstYearGuide/mobile/assets/css/components/road.css
@@ -1,3 +1,0 @@
-/* .road-image {
-    width: auto;
-} */
