+++ src/firstYearGuide/mobile/assets/css/mobileView.css
@@ -67,22 +67,7 @@ body {
     z-index: 3;
 }
 
-.road-container {
-    position: absolute;
-    bottom: 0;
-    left: 0;
 
-    width: 100%;
-    height: 100%;
-
-    background-image: url("../images/components/mobile-road.png");
-    background-repeat: repeat-x;
-    background-size: auto 100%;
-
-    overflow: hidden;
-
-    z-index: 2;
-}
 
 .starsContainer {
     width: 600vw;
