--- src/firstYearGuide/mobile/components/road/road.jsx
@@ -1,8 +1,0 @@
-// import roadImage from '../../assets/images/components/mobile-road.png';
-// import '../../assets/css/components/road.css';
-
-// const Road = () => (
-//   <img src={roadImage} alt="Road" className="road-image" />
-// );
-
-// export default Road;
