--- src/firstYearGuide/mobile/components/scrollButton/scrollButton.css
@@ -1,20 +1,0 @@
-.scrollButton {
-    margin-left: 10%;
-    margin-right: 10%;
-    
-    height: 100%;
-    width: 100%;
-
-    border: none;
-    background-color: transparent;
-    background-repeat: no-repeat;
-    background-size: cover;
-}
-
-.scrollButtonLeft{
-    background-image: url('./assets/mobile-Arrow-left.svg');
-}
-
-.scrollButtonRight{
-    background-image: url('./assets/mobile-Arrow-right.svg');
-}
